+++ demo/run.sh
@@ -21,10 +21,9 @@ echo "keygen part"
 
 for i in $(seq 1 $n)
 do
-
-echo "key gen for client $i out of $n"
-./target/release/examples/gg18_keygen_client http://127.0.0.1:8001 keys$i.store &
-sleep 3
+    echo "key gen for client $i out of $n"
+    ./target/release/examples/gg18_keygen_client http://127.0.0.1:8001 keys$i.store &
+    sleep 3
 done
 
 
@@ -34,11 +33,9 @@ echo "sign"
 
 for i in $(seq 1 $((t+1)));
 do
-echo "signing for client $i out of $((t+1))"
-./target/release/examples/gg18_sign_client http://127.0.0.1:8001 keys$i.store "KZen Networks" &
-
-sleep 2
+    echo "signing for client $i out of $((t+1))"
+    ./target/release/examples/gg18_sign_client http://127.0.0.1:8001 keys$i.store "KZen Networks" &
+    sleep 3
 done
 
-rm params.json
-killall sm_manager gg18_keygen_client gg18_sign_client 2> /dev/null
+killall sm_manager 2> /dev/null
